new order of components, separating by pages and components
