+++ src/components/petProject/home/HomeBody.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { Button } from "../ui/button";
-import { Label } from "../ui/label";
+import { Button } from "../../ui/button";
+import { Label } from "../../ui/label";
 
 const HomeBodyComponent: React.FC = () => {
   return (
