+++ src/components/petProject/home/HomeMotivationalMessage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Button } from "../ui/button";
-import { Card } from "../ui/card";
-import { Label } from "../ui/label";
+import { Button } from "../../ui/button";
+import { Card } from "../../ui/card";
+import { Label } from "../../ui/label";
 
 const HomeMotivationalMessageComponent: React.FC = () => {
   return (
