+++ src/components/petProject/home/HomeProcess.tsx
@@ -1,7 +1,7 @@
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
 import { ArrowRight } from "lucide-react";
 import React from "react";
-import { Input } from "../ui/input";
-import { Label } from "../ui/label";
 import HomeProcessStep from "./HomeProcessStep";
 
 const HomeProcessComponent: React.FC = () => {
