+++ src/shared/Footer.tsx
@@ -1,7 +1,7 @@
 import { ArrowRight } from "lucide-react";
 import React from "react";
-import { Input } from "../ui/input";
-import { Label } from "../ui/label";
+import { Input } from "../components/ui/input";
+import { Label } from "../components/ui/label";
 
 const FooterComponent: React.FC = () => {
   return (
