+++ src/shared/HomeNavbar.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Label } from "../ui/label";
+import { Label } from "../components/ui/label";
 
 const HomeNavbarComponent: React.FC = () => {
   return (
